Rename files in common/QueueStats
